Improve FreeRDP authentication failure handling
The issue is caused by changes introduced by 60dea27. When the RDP
server refuses TCP connections on the specified port, freerdp_connect()
causes freerdp_get_last_error() to return 0x2000c.  Due to the way the
do-while loop is constructed in open_freerdp(), this effectively causes
freerdp_connect() to be called in an infinite loop, which quickly
results in a SIGSEGV as FreeRDP does not seem to handle eventfd
creation failures nicely (these are caused by exceeding the per-process
limit of open file descriptors).

As the error string for error 0x2000c is "protocol security negotiation
or connection failure", to at least allow refused connections to be
handled gracefully, some limit should be enforced on the number of loop
iterations inside open_freerdp().

https://bugzilla.gnome.org/show_bug.cgi?id=769217